Implements a 90% working Editor.
The ClassItem definition was updated to support more getters, and setters for all it's options. This way, it can trigger an onChange event when the editor updates the ClassItem.
Most of the Editors interface is working, including adding Interfaces, Members, and Methods, as well as editing their names, values, and properties in their tables. Of course the basics work, such as isPublic, isFinal, isAbstract and class name / ancestor name.

The primary things the Editor needs are:
- Ability to remove items from the table lists
- Ability to set initial values to be "unset"
- Ability for it to auto correct the value based on type
- Abilty to Add / Edit / Remove parameters for methods