Updates the editor with the ability to delete rows, and have unset initial member values.
Now rows can be deleted, and member variables can have an unset value.
This commit also provides a basic implementation of value sanitization for member variables.
It will correctly sanitize them if the type is changed, but the interface wont update right away.
Need a smarter way for it to know that the value has changed.