Implements the default set of language generators, up & running
The generators for: Java, C#, Python, Ruby, PHP, JavaScript (Vanilla, ES6, TypeScript) and VB (.NET, 6) are all implement and more-or-less generating correct code. Python could use some fixing up according to Ahmed. Most of the warning generators do not work yet, and most could use a little refactoring.

The editor fixes some glitches with data entry, and a file of default-classes was created so the editor would be able to be populated with default examples.